VirtueIR  creating our application using Vite and installing and setting up tailwind css
Navbar which is fully responsive and has a mobile drawer to display all links of the page.
The website also contains a beautiful hero section which uses gradient text, to give a modern look to the site. We will also be using video tag to display few videos in autoplay and loop mode. The site also contains a feature section that lists all the features of the product displayed in a clean UI.
The site will also have a pricing card feature with three plans for the users to select. We will also look at how to create testimonials and footer section using tailwind css.